+++ retail/BetterBlizzFrames.lua
@@ -2,7 +2,7 @@
 
 local addonVersion = "1.00" --too afraid to to touch for now
 local addonUpdates = C_AddOns.GetAddOnMetadata("BetterBlizzFrames", "Version")
-local sendUpdate = true
+local sendUpdate = false
 BBF.VersionNumber = addonUpdates
 BBF.variablesLoaded = false
 local isAddonLoaded = C_AddOns.IsAddOnLoaded
@@ -3752,6 +3752,7 @@ Frame:SetScript("OnEvent", function(...)
             end
             BBF.DarkmodeFrames()
             BBF.HookPlayerAndTargetAuras()
+            BBF.HookFrameTextureColor()
 
             if BetterBlizzFramesDB.playerFrameOCD then
                 BBF.FixStupidBlizzPTRShit()